pyproject.toml
[project]
name = "project-softdesk-v2"
version = "0.1.0"
description = ""
authors = [
    {name = "meridien22",email = "meridien.sig@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]

[tool.poetry]
packages = [{include = "project_softdesk_v2", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]
